+++ backend/src/main/java/com/ssafy/jupging/service/ArticleService.java
@@ -22,7 +22,7 @@ public class ArticleService {
     }
 
     @Transactional
-    public Article findByArticleId(int article_id) {
+    public Article findByArticleId(Long article_id) {
         Article article = articleRepository.findByArticleId(article_id)
                 .orElseThrow(()->new IllegalArgumentException("게시글 정보가 없습니다"));
